refactor usage of transactionConfirmed
